requirements.txt
asgiref==3.9.1
Django==5.2.4
django-ckeditor==6.7.3
django-extensions==4.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-taggit==6.1.0
MarkupSafe==3.0.2
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
whitenoise==6.9.0
